@@ -12,7 +12,7 @@
     "posttest": "yarn run lint",
     "serve": "webpack-dev-server --env=dev",
     "serve:dist": "webpack-dev-server --env=dist",
-    "start": "webpack-dev-server --env=dev",
+    "start": "webpack-dev-server --env=dev --hot --mode development",
     "test": "karma start",
     "test:watch": "karma start --autoWatch=true --singleRun=false"
   },
@@ -21,68 +21,70 @@
   "author": "Will James",
   "devDependencies": {
     "babel-core": "^6.0.0",
-    "babel-eslint": "^6.0.0",
-    "babel-loader": "^6.0.0",
+    "babel-eslint": "^8.2.2",
+    "babel-loader": "^7.1.3",
     "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.0.15",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.0.15",
-    "chai": "^3.2.0",
+    "chai": "^4.1.2",
     "copyfiles": "^1.0.0",
-    "css-loader": "^0.23.0",
-    "eslint": "^3.0.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^5.0.0",
+    "css-loader": "^0.28.10",
+    "eslint": "^4.18.1",
+    "eslint-loader": "^2.0.0",
+    "eslint-plugin-react": "^7.7.0",
     "explorer-cdn": "https://github.com/sportnumerics/explorer-cdn#1.0.1",
-    "file-loader": "^0.9.0",
+    "file-loader": "^1.1.10",
     "glob": "^7.0.0",
-    "html-webpack-inline-source-plugin": "0.0.8",
-    "html-webpack-plugin": "^2.28.0",
-    "istanbul-instrumenter-loader": "^2.0.0",
+    "html-webpack-inline-source-plugin": "0.0.9",
+    "html-webpack-plugin": "^3.0.4",
+    "istanbul-instrumenter-loader": "^3.0.0",
     "jimp": "^0.2.27",
-    "karma": "^1.0.0",
+    "karma": "^2.0.0",
     "karma-chai": "^0.1.0",
     "karma-coverage": "^1.0.0",
     "karma-mocha": "^1.0.0",
     "karma-mocha-reporter": "^2.0.0",
     "karma-sourcemap-loader": "^0.3.5",
-    "karma-webpack": "^1.7.5",
+    "karma-webpack": "^2.0.13",
     "minimist": "^1.2.0",
-    "mocha": "^3.0.0",
+    "mocha": "^5.0.1",
     "node-sass": "^4.7.2",
     "null-loader": "^0.1.1",
     "open": "0.0.5",
-    "postcss": "^5.0.11",
-    "postcss-loader": "^0.8.0",
+    "postcss": "^6.0.19",
+    "postcss-loader": "^2.1.1",
     "react-addons-test-utils": "^15.0.0",
-    "react-hot-loader": "^1.3.1",
-    "responsive-loader": "^0.7.0",
+    "react-hot-loader": "^4.0.0",
+    "responsive-loader": "^1.1.0",
     "rimraf": "^2.4.3",
     "sass-loader": "^6.0.2",
-    "style-loader": "^0.13.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.11.1"
+    "style-loader": "^0.20.2",
+    "url-loader": "^0.6.2",
+    "webpack": "^4.0.1",
+    "webpack-cli": "^2.0.9",
+    "webpack-dev-server": "^3.1.0"
   },
   "dependencies": {
     "bootstrap-sass": "^3.3.7",
     "core-js": "^2.0.0",
-    "es6-promise": "^3.2.1",
+    "es6-promise": "^4.2.4",
     "font-awesome": "^4.7.0",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "https://registry.npmjs.org/lodash/-/lodash-4.17.4.tgz",
     "moment": "^2.17.1",
-    "normalize.css": "^4.0.0",
-    "react": "^15.0.0",
-    "react-bootstrap": "^0.30.8",
-    "react-dom": "^15.0.0",
+    "normalize.css": "^8.0.0",
+    "react": "^16.2.0",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^16.2.0",
     "react-fontawesome": "^1.1.0",
     "react-ga": "^2.1.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^3.0.2",
-    "react-router-bootstrap": "^0.23.1",
+    "react-redux": "^5.0.7",
+    "react-router": "^4.2.0",
+    "react-router-bootstrap": "^0.24.4",
+    "react-router-dom": "^4.2.2",
     "redux": "^3.5.2",
-    "redux-actions": "^0.10.1",
-    "redux-logger": "^2.6.1",
+    "redux-actions": "^2.2.1",
+    "redux-logger": "^3.0.6",
     "redux-promise": "^0.5.3",
     "redux-thunk": "^2.1.0"
   }
