@@ -4,7 +4,7 @@ import baseConfig from './base';
 
 let config = {
   appEnv: 'dev',  // feel free to remove the appEnv property here
-  apiUrl: 'https://explorer-api.dev.sportnumerics.com'
+  apiUrl: 'https://explorer-api-blue.sportnumerics.com'
 };
 
 export default Object.freeze(Object.assign({}, baseConfig, config));
