@@ -2,13 +2,13 @@ import 'core-js/fn/object/assign';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Root from './components/Root';
-import { browserHistory } from 'react-router';
+import history from './services/history';
 import ReactGA from 'react-ga';
 import config from 'config';
 
 ReactGA.initialize(config.gaTrackingId, config.gaOptions);
 
-browserHistory.listen( location =>  {
+history.listen( location =>  {
   ReactGA.set({ page: location.pathname });
   ReactGA.pageview(location.pathname);
 });
